pyproject.toml
[tool]
[tool.poetry]
name = "imagery"
version = "0.1.0"
description = "A CMS for artists."
license = "GPL-3.0-or-later"
authors = ["Jan Willems <jw@elevenbits.com>"]
maintainers = ["Jan Willems <jw@elevenbits.com>"]
readme = "README.rst"
homepage = "https://github.com/jw/stilus"
repository = "https://github.com/jw/stilus"
documentation = "https://stilus.readthedocs.io/en/latest/"
keywords = ["compiler", "stylus", "css"]
classifiers = [
    'License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: Implementation :: CPython'
]
include = ['README.md', 'LICENSE']

[tool.poetry.dependencies]
python = "==3.7.7"
awesome-slugify = "==1.6.5"
boto = "==2.42.0"
boto3 = "*"
django = "==2.1"
django-allauth = "==0.36.0"
django-anymail = "==3.0"
django-braces = "==1.13.0"
django-ckeditor = "*"
django-crispy-forms = "==1.7.2"
django-debug-toolbar = "*"
django-environ = "==0.4.5"
django-extensions = "==2.1.0"
django-markdownx = "*"
django-model-utils = "==3.1.2"
django-redis = "==4.5.0"
django-sitetree = "==1.11.0"
django-storages = "*"
gevent = "==1.3.5"
gunicorn = "==19.9.0"
pillow = "==5.2.0"
psycopg2 = "==2.7.4"
pytz = "==2016.7"
raven = "==6.9.0"
redis = ">=2.10.5"
whitenoise = "==3.3.1"

[tool.poetry.dev-dependencies]
ipython = "*"
werkzeug = "*"
pre-commit = "^2.7.1"
black = "^20.8b1"
flake8 = "^3.8.3"
